refactor config to standalone and use in ./index.js
Config should be loaded in `./index.js` instead of `./bin/index.js` otherwise modules using `require('lighthouse-security')` don't get the security specific config. Fixes #9.

Config is moved to `./config` uses all absolute paths and no longer extends `'lighthouse:default'` so it can be used standalone and other modules can use it using `require('lighthouse-security/config')`. Extension is now done where the config is actually used, in `./index.js`.

Google license is removed from `./config.js` as the file is not made Google. Fixes #3.
