+++ index.js
@@ -1,20 +1,15 @@
-const lighthouse = require('lighthouse');
-const chromeLauncher = require('lighthouse/chrome-launcher');
+const lighthouse = require('lighthouse')
+const chromeLauncher = require('lighthouse/chrome-launcher')
+const pageSecurityConfig = require('./config')
 
-async function run(url, flags = {}, config = null) {
+async function run(url, flags = {}) {
+    const config = Object.assign({ extends: 'lighthouse:default' }, pageSecurityConfig)
     const chrome = await chromeLauncher.launch()
     flags.port = chrome.port
     const results = await lighthouse(url, flags, config)
     const stopped = await chrome.kill()
 
     return results
-    //
-    // return chromeLauncher.launch()
-    //     .then(chrome => {
-    //         flags.port = chrome.port;
-    //         return lighthouse(url, flags, config)
-    //             .then(results => chrome.kill().then(() => results));
-    //     });
 }
 
 module.exports = run
